refactor: uses export instead export default
