+++ src/App.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import './App.css'
 
 import Routes from './routes'
